refactor(paginate): Group tests into their own files
GTC-1865
