+++ tests_v2/unit/app/paginate/test_pagination_meta_info.py
@@ -3,69 +3,7 @@ from unittest.mock import Mock
 import pytest
 
 from app.crud.datasets import count_datasets, get_datasets
-from app.models.orm.datasets import Dataset as ORMDataset
 from app.paginate.paginate import PaginationMeta, paginate_datasets
-
-
-@pytest.mark.asyncio
-async def test_legacy_no_pagination_happens_for_default_values():
-    """This is for legacy compatibility."""
-    spy_get_datasets = Mock(get_datasets)
-
-    await paginate_datasets(crud_impl=spy_get_datasets)
-
-    spy_get_datasets.assert_called_with(None, 0)
-
-
-@pytest.mark.asyncio
-async def test_offset_is_0_for_page_1_when_size_is_given():
-    spy_get_datasets = Mock(get_datasets)
-    dummy_count_datasets = Mock(count_datasets)
-
-    await paginate_datasets(crud_impl=spy_get_datasets, datasets_count_impl=dummy_count_datasets, size=10, page=1)
-
-    spy_get_datasets.assert_called_with(size=10, offset=0)
-
-
-@pytest.mark.asyncio
-async def test_offset_is_0_when_no_page_is_given():
-    spy_get_datasets = Mock(get_datasets)
-    dummy_count_datasets = Mock(count_datasets)
-
-    await paginate_datasets(crud_impl=spy_get_datasets, datasets_count_impl=dummy_count_datasets, size=10)
-
-    spy_get_datasets.assert_called_with(size=10, offset=0)
-
-
-@pytest.mark.asyncio
-async def test_offset_is_10_for_page_2_when_page_size_is_10():
-    spy_get_datasets = Mock(get_datasets)
-    dummy_count_datasets = Mock(count_datasets)
-
-    await paginate_datasets(crud_impl=spy_get_datasets, datasets_count_impl=dummy_count_datasets, size=10, page=2)
-
-    spy_get_datasets.assert_called_with(size=10, offset=10)
-
-
-@pytest.mark.asyncio
-async def test_legacy_datasets_collection_is_returned_when_no_arguments_are_given():
-    """This is for legacy compatibility."""
-    stub_get_datasets = Mock(get_datasets)
-    stub_get_datasets.return_value = [ORMDataset()]
-
-    data, _ = await paginate_datasets(crud_impl=stub_get_datasets)
-
-    assert isinstance(data, list)
-    assert isinstance(data[0], ORMDataset)
-
-
-@pytest.mark.asyncio
-async def test_legacy_meta_section_is_none_when_no_arguments_are_given():
-    dummy_get_datasets = Mock(get_datasets)
-
-    _, meta = await paginate_datasets(crud_impl=dummy_get_datasets)
-
-    assert meta is None
 
 
 @pytest.mark.asyncio
